Remove dependency on Regex
Until we can begin mirroring our Cargo dependencies internally, we are
going to want to vendor our dependencies so that we don't add another
external site as a dependency of our build. In order to minimize the
amount of third party code that we have to check in, we want to minimize
our set of dependencies by removing ones that aren't strictly necessary.
One crate which we can easily remove is Regex.